requirements.txt
flask==1.1.2
mysql-connector-python==8.0.21
tappay==0.4.0
python-dotenv==0.13.0